+++ .gitignore
@@ -194,3 +194,4 @@ $RECYCLE.BIN/
 # Backend Ignore
 
 backend/static
+backend/generated
