+++ backend/package.json
@@ -27,6 +27,7 @@
     "@nestjs/passport": "^10.0.3",
     "@nestjs/platform-express": "^10.0.0",
     "@nestjs/swagger": "^7.1.14",
+    "@prisma/client": "^5.9.1",
     "@supabase/supabase-js": "^2.39.3",
     "class-transformer": "^0.5.1",
     "class-validator": "^0.14.0",
